move reading of the test file to outside the tester package
