+++ go/tester/comparing_query_runner.go
@@ -18,6 +18,7 @@ package tester
 
 import (
 	"fmt"
+	"github.com/vitessio/vitess-tester/go/tools"
 
 	"github.com/pingcap/errors"
 	log "github.com/sirupsen/logrus"
@@ -54,11 +55,11 @@ func newComparingQueryRunner(
 	}
 }
 
-func (nqr ComparingQueryRunner) runQuery(q query, expectedErrs bool, ast sqlparser.Statement) error {
+func (nqr ComparingQueryRunner) runQuery(q tools.Query, expectedErrs bool, ast sqlparser.Statement) error {
 	return nqr.execute(q, expectedErrs, ast)
 }
 
-func (nqr *ComparingQueryRunner) execute(query query, expectedErrs bool, ast sqlparser.Statement) error {
+func (nqr *ComparingQueryRunner) execute(query tools.Query, expectedErrs bool, ast sqlparser.Statement) error {
 	if len(query.Query) == 0 {
 		return nil
 	}
